port over grade tosdr logic to new singleton
