+++ shared/js/background/classes/score.es6.js
@@ -1,9 +1,6 @@
 const tldjs = require('tldjs')
-const tosdr = require('../../../data/tosdr')
 const constants = require('../../../data/constants')
 const utils = require('../utils.es6')
-const tosdrRegexList = Object.keys(tosdr).map(x => new RegExp(`(^)${tldjs.getDomain(x)}`)) // only match domains, and from the start of the URL
-const tosdrClassMap = {'A': -1, 'B': 0, 'C': 0, 'D': 1, 'E': 2} // map tosdr class rankings to increase/decrease in grade
 const siteScores = ['A', 'B', 'C', 'D']
 const pagesSeenOn = constants.majorTrackingNetworks
 
@@ -16,58 +13,8 @@ class Score {
         this.hasObscureTracker = false
         this.domain = tldjs.getDomain(domain) // strip the subdomain. Fixes matching tosdr for eg encrypted.google.com
         this.isaMajorTrackingNetwork = this.isaMajorTrackingNetwork()
-        this.tosdr = this.getTosdr()
     }
 
-    getTosdr () {
-        let result = {}
-
-        tosdrRegexList.some(tosdrSite => {
-            let match = tosdrSite.exec(this.domain)
-            if (match) {
-                // remove period at end for lookup in pagesSeenOn
-                let tosdrData = tosdr[match[0]]
-
-                if (!tosdrData) return
-
-                const matchGood = (tosdrData.match && tosdrData.match.good) || []
-                const matchBad = (tosdrData.match && tosdrData.match.bad) || []
-
-                // tosdr message
-                // 1. If we have a defined tosdr class look up the message in constants
-                //    for the corresponding letter class
-                // 2. If there are both good and bad points -> 'mixed'
-                // 3. Else use the calculated tosdr score to determine the message
-                let message = constants.tosdrMessages.unknown
-                if (tosdrData.class) {
-                    message = constants.tosdrMessages[tosdrData.class]
-                } else if (matchGood.length && matchBad.length) {
-                    message = constants.tosdrMessages.mixed
-                } else {
-                    if (tosdrData.score < 0) {
-                        message = constants.tosdrMessages.good
-                    } else if (tosdrData.score === 0 && (matchGood.length || matchBad.length)) {
-                        message = constants.tosdrMessages.mixed
-                    } else if (tosdrData.score > 0) {
-                        message = constants.tosdrMessages.bad
-                    }
-                }
-
-                result = {
-                    score: tosdrData.score,
-                    class: tosdrData.class,
-                    reasons: {
-                        good: matchGood,
-                        bad: matchBad
-                    },
-                    message: message
-                }
-
-                return result
-            }
-        })
-        return result
-    }
 
     /* is the parent site itself a major tracking network?
      * minus one grade for each 10% of the top pages this
@@ -99,20 +46,6 @@ class Score {
             afterIndex += this.isaMajorTrackingNetwork
         }
 
-        // If tosdr already determined a class ranking then we map that to increase or
-        // decrease the grade accordingly. Otherwise we apply a +/- to the grade based
-        // on the cumulative total of all the points we care about. see: scripts/tosdr-topics.json
-        if (this.tosdr) {
-            if (this.tosdr.class) {
-                beforeIndex += tosdrClassMap[this.tosdr.class]
-                afterIndex += tosdrClassMap[this.tosdr.class]
-            } else if (this.tosdr.score) {
-                let tosdrScore = Math.sign(this.tosdr.score)
-                beforeIndex += tosdrScore
-                afterIndex += tosdrScore
-            }
-        }
-
         if (this.inMajorTrackingNetwork) beforeIndex++
         if (!this.hasHTTPS) {
             beforeIndex++
@@ -129,8 +62,8 @@ class Score {
         if (afterIndex < 0) afterIndex = 0
 
         // only sites with a tosdr.class "A" can get a final grade of "A"
-        if (afterIndex === 0 && this.tosdr.class !== 'A') afterIndex = 1
-        if (beforeIndex === 0 && this.tosdr.class !== 'A') beforeIndex = 1
+        // if (afterIndex === 0 && this.tosdr.class !== 'A') afterIndex = 1
+        // if (beforeIndex === 0 && this.tosdr.class !== 'A') beforeIndex = 1
 
         // return corresponding score or lowest score if outside the array
         let beforeGrade = siteScores[beforeIndex] || siteScores[siteScores.length - 1]
